refactor(extension): extract reporting functions to dedicated module
