+++ vscode-gotest/src/testController.ts
@@ -330,129 +330,6 @@ export class GoTestController implements vscode.Disposable {
     await this.resultStore.save();
   }
 
-  async copyTestResults(rootItem?: vscode.TestItem): Promise<void> {
-    type Agg = {
-      passed: number;
-      failed: number;
-      skipped: number;
-      duration: number;
-    };
-    type Row = {
-      label: string;
-      duration?: number;
-      status?: string;
-      agg?: Agg;
-    };
-    const rows: Row[] = [];
-
-    const walkItem = (item: vscode.TestItem, indent: number): Agg => {
-      const structural =
-        item.id.startsWith("dir:") || item.tags.some((t) => t.id === "package");
-      const result = structural ? undefined : this.resultStore.get(item.id);
-
-      const rowIdx = rows.length;
-      rows.push({
-        label: "  ".repeat(indent) + item.label,
-        duration: result?.duration,
-        status: result?.status,
-      });
-
-      const childAgg: Agg = { passed: 0, failed: 0, skipped: 0, duration: 0 };
-      item.children.forEach((child) => {
-        const ca = walkItem(child, indent + 1);
-        childAgg.passed += ca.passed;
-        childAgg.failed += ca.failed;
-        childAgg.skipped += ca.skipped;
-        childAgg.duration += ca.duration;
-      });
-
-      if (item.children.size > 0) {
-        rows[rowIdx].agg = childAgg;
-        return childAgg;
-      }
-
-      const leafAgg: Agg = { passed: 0, failed: 0, skipped: 0, duration: 0 };
-      if (result?.status === "pass") leafAgg.passed = 1;
-      else if (result?.status === "fail") leafAgg.failed = 1;
-      else if (result?.status === "skip") leafAgg.skipped = 1;
-      if (result?.duration) leafAgg.duration = result.duration;
-      return leafAgg;
-    };
-
-    const resolved = rootItem ? this.findItem(rootItem.id) : undefined;
-    if (resolved) {
-      walkItem(resolved, 0);
-    } else {
-      this.controller.items.forEach((item) => walkItem(item, 0));
-    }
-
-    if (rows.length === 0) {
-      vscode.window.showInformationMessage(
-        "No test items available. Run discovery first.",
-      );
-      return;
-    }
-
-    const maxLabelLen = Math.max(4, ...rows.map((r) => r.label.length));
-    const header = `${"Test".padEnd(maxLabelLen)}  Time       Result`;
-    const separator = "-".repeat(header.length);
-
-    const lines = [header, separator];
-    let totalPassed = 0;
-    let totalFailed = 0;
-    let totalSkipped = 0;
-    let totalDuration = 0;
-
-    for (const row of rows) {
-      if (row.agg) {
-        const a = row.agg;
-        const aggTime =
-          a.duration > 0 ? (a.duration / 1000).toFixed(3) + "s" : "-";
-        const parts: string[] = [];
-        if (a.passed > 0) parts.push(`${a.passed} passed`);
-        if (a.failed > 0) parts.push(`${a.failed} failed`);
-        if (a.skipped > 0) parts.push(`${a.skipped} skipped`);
-        const aggSummary = parts.length > 0 ? parts.join(", ") : "-";
-        lines.push(
-          `${row.label.padEnd(maxLabelLen)}  ${aggTime.padEnd(9)}  ${aggSummary}`,
-        );
-        continue;
-      }
-
-      const time =
-        row.duration !== undefined
-          ? (row.duration / 1000).toFixed(3) + "s"
-          : "-";
-      const status = row.status ?? "-";
-      lines.push(
-        `${row.label.padEnd(maxLabelLen)}  ${time.padEnd(9)}  ${status}`,
-      );
-
-      if (row.status === "pass") totalPassed++;
-      else if (row.status === "fail") totalFailed++;
-      else if (row.status === "skip") totalSkipped++;
-      if (row.duration) totalDuration += row.duration;
-    }
-
-    lines.push(separator);
-    const hasResults = totalPassed + totalFailed + totalSkipped > 0;
-    if (hasResults) {
-      const parts: string[] = [];
-      if (totalPassed > 0) parts.push(`${totalPassed} passed`);
-      if (totalFailed > 0) parts.push(`${totalFailed} failed`);
-      if (totalSkipped > 0) parts.push(`${totalSkipped} skipped`);
-      lines.push(
-        `Total: ${parts.join(", ")} (${(totalDuration / 1000).toFixed(3)}s)`,
-      );
-    } else {
-      lines.push("Total: no results");
-    }
-
-    const text = lines.join("\n");
-    await vscode.env.clipboard.writeText(text);
-    vscode.window.showInformationMessage("Test results copied to clipboard.");
-  }
-
   dispose(): void {
     for (const d of this.disposables) {
       d.dispose();
